Revert "Cruved - Prise en compte de l'héritage objet, module, GEONATURE"
